Split files for service interface and impl
This is just a quality-of-life change to clarify intra-package
dependencies.

The option types accepted by the service interface were simplified to
internally cast to a single interface rather than every possible
struct, localizing changes where such structs are defined rather than
scattering them into every `WithXXX` function.

The service implementation was just split into multiple files with the
idea of clarifying which interface methods depends on which internal
routines.

Finally, I replaced all `interface{}` with `any` because the IDE was
annoying me.
