+++ internal/api/common/common.go
@@ -7,7 +7,7 @@ import (
 )
 
 // WriteJSONResponse writes a JSON response with the given data
-func WriteJSONResponse(w http.ResponseWriter, data interface{}, statusCode int) {
+func WriteJSONResponse(w http.ResponseWriter, data any, statusCode int) {
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(statusCode)
 	if err := json.NewEncoder(w).Encode(data); err != nil {
