+++ internal/registry/tags_test.go
@@ -34,9 +34,9 @@ func TestExtractTags(t *testing.T) {
 			name: "test with tags",
 			server: &upstream.ServerJSON{
 				Meta: &upstream.ServerMeta{
-					PublisherProvided: map[string]interface{}{
-						"provider": map[string]interface{}{
-							"image": map[string]interface{}{"tags": []interface{}{"tag1", "tag2"}},
+					PublisherProvided: map[string]any{
+						"provider": map[string]any{
+							"image": map[string]any{"tags": []any{"tag1", "tag2"}},
 						},
 					},
 				},
@@ -47,7 +47,7 @@ func TestExtractTags(t *testing.T) {
 			name: "test without tags",
 			server: &upstream.ServerJSON{
 				Meta: &upstream.ServerMeta{
-					PublisherProvided: map[string]interface{}{},
+					PublisherProvided: map[string]any{},
 				},
 			},
 			expectedTags: []string{},
@@ -61,13 +61,13 @@ func TestExtractTags(t *testing.T) {
 			name: "test with string values in PublisherProvided (issue #299)",
 			server: &upstream.ServerJSON{
 				Meta: &upstream.ServerMeta{
-					PublisherProvided: map[string]interface{}{
+					PublisherProvided: map[string]any{
 						// String value that would cause panic without fix
 						"someKey": "someStringValue",
 						// Map value with tags (expected structure)
-						"provider": map[string]interface{}{
-							"metadata": map[string]interface{}{
-								"tags": []interface{}{"tag1", "tag2"},
+						"provider": map[string]any{
+							"metadata": map[string]any{
+								"tags": []any{"tag1", "tag2"},
 							},
 						},
 					},
