+++ internal/service/service.go
@@ -4,7 +4,6 @@ package service
 import (
 	"context"
 	"errors"
-	"fmt"
 	"time"
 
 	upstreamv0 "github.com/modelcontextprotocol/registry/pkg/api/v0"
@@ -53,11 +52,16 @@ var (
 
 // RegistryService defines the interface for registry operations
 type RegistryService interface {
+
+	// ********** COMMON OPERATIONS **********
+
 	// CheckReadiness checks if the regSvc is ready to serve requests
 	CheckReadiness(ctx context.Context) error
 
 	// GetRegistry returns the registry data with metadata
 	GetRegistry(ctx context.Context) (*toolhivetypes.UpstreamRegistry, string, error) // returns registry, source, error
+
+	// ********** MCP OPERATIONS **********
 
 	// ListServers returns all servers in the registry with pagination info
 	ListServers(ctx context.Context, opts ...Option) (*ListServersResult, error)
@@ -73,6 +77,8 @@ type RegistryService interface {
 
 	// DeleteServerVersion removes a server version from a managed registry
 	DeleteServerVersion(ctx context.Context, opts ...Option) error
+
+	// ********** REGISTRY OPERATIONS **********
 
 	// ListRegistries returns all configured registries
 	ListRegistries(ctx context.Context) ([]RegistryInfo, error)
@@ -91,6 +97,8 @@ type RegistryService interface {
 
 	// ProcessInlineRegistryData processes inline data for a managed/file registry
 	ProcessInlineRegistryData(ctx context.Context, name string, data string, format string) error
+
+	// ********** SKILL OPERATIONS **********
 
 	// ListSkills lists skills in a registry with cursor-based pagination
 	ListSkills(ctx context.Context, opts ...Option) (*ListSkillsResult, error)
@@ -113,7 +121,7 @@ type RegistryInfo struct {
 	CreationType CreationType         `json:"creationType,omitempty"` // API or CONFIG
 	SourceType   config.SourceType    `json:"sourceType,omitempty"`   // git, api, file, managed, kubernetes
 	Format       string               `json:"format,omitempty"`       // toolhive or upstream
-	SourceConfig interface{}          `json:"sourceConfig,omitempty"` // Type-specific source configuration
+	SourceConfig any                  `json:"sourceConfig,omitempty"` // Type-specific source configuration
 	FilterConfig *config.FilterConfig `json:"filterConfig,omitempty"` // Filtering rules
 	SyncSchedule string               `json:"syncSchedule,omitempty"` // Sync interval string
 	SyncStatus   *RegistrySyncStatus  `json:"syncStatus,omitempty"`
@@ -134,297 +142,6 @@ type RegistrySyncStatus struct {
 // RegistryListResponse represents the response for listing registries
 type RegistryListResponse struct {
 	Registries []RegistryInfo `json:"registries"`
-}
-
-// Option is a function that sets an option for the ListServersOptions, ListServerVersionsOptions,
-// GetServerVersionOptions, PublishServerVersionOptions, or DeleteServerVersionOptions
-type Option func(any) error
-
-// ListServersOptions is the options for the ListServers operation
-type ListServersOptions struct {
-	RegistryName *string
-	Cursor       string
-	Limit        int
-	Search       string
-	UpdatedSince time.Time
-	Version      string
-}
-
-// ListServerVersionsOptions is the options for the ListServerVersions operation
-type ListServerVersionsOptions struct {
-	RegistryName *string
-	Name         string
-	Next         *time.Time
-	Prev         *time.Time
-	Limit        int
-}
-
-// GetServerVersionOptions is the options for the GetServerVersion operation
-type GetServerVersionOptions struct {
-	RegistryName string
-	Name         string
-	Version      string
-}
-
-// PublishServerVersionOptions is the options for the PublishServerVersion operation
-type PublishServerVersionOptions struct {
-	RegistryName string
-	ServerData   *upstreamv0.ServerJSON
-}
-
-// DeleteServerVersionOptions is the options for the DeleteServerVersion operation
-type DeleteServerVersionOptions struct {
-	RegistryName string
-	ServerName   string
-	Version      string
-}
-
-// WithCursor sets the cursor for the ListServers operation
-func WithCursor(cursor string) Option {
-	return func(o any) error {
-		if cursor == "" {
-			return fmt.Errorf("invalid cursor: %s", cursor)
-		}
-
-		switch o := o.(type) {
-		case *ListServersOptions:
-			o.Cursor = cursor
-		case *ListSkillsOptions:
-			o.Cursor = &cursor
-		default:
-			return fmt.Errorf("invalid option type: %T", o)
-		}
-
-		return nil
-	}
-}
-
-// WithSearch sets the search for the ListServers operation
-func WithSearch(search string) Option {
-	return func(o any) error {
-		if search == "" {
-			return fmt.Errorf("invalid search: %s", search)
-		}
-
-		switch o := o.(type) {
-		case *ListServersOptions:
-			o.Search = search
-		case *ListSkillsOptions:
-			o.Search = &search
-		default:
-			return fmt.Errorf("invalid option type: %T", o)
-		}
-
-		return nil
-	}
-}
-
-// WithUpdatedSince sets the updated since for the ListServers operation
-func WithUpdatedSince(updatedSince time.Time) Option {
-	return func(o any) error {
-		if updatedSince.IsZero() {
-			return fmt.Errorf("invalid updated since: %s", updatedSince)
-		}
-
-		switch o := o.(type) {
-		case *ListServersOptions:
-			o.UpdatedSince = updatedSince
-		default:
-			return fmt.Errorf("invalid option type: %T", o)
-		}
-
-		return nil
-	}
-}
-
-// WithRegistryName sets the registry name for the ListServers, ListServerVersions,
-// GetServerVersion, PublishServerVersion, or DeleteServerVersion operation
-func WithRegistryName(registryName string) Option {
-	return func(o any) error {
-		if registryName == "" {
-			return fmt.Errorf("invalid registry name: %s", registryName)
-		}
-
-		switch o := o.(type) {
-		case *ListServersOptions:
-			o.RegistryName = &registryName
-		case *ListServerVersionsOptions:
-			o.RegistryName = &registryName
-		case *GetServerVersionOptions:
-			o.RegistryName = registryName
-		case *PublishServerVersionOptions:
-			o.RegistryName = registryName
-		case *DeleteServerVersionOptions:
-			o.RegistryName = registryName
-		case *ListSkillsOptions:
-			o.RegistryName = registryName
-		case *GetSkillVersionOptions:
-			o.RegistryName = registryName
-		case *PublishSkillOptions:
-			o.RegistryName = registryName
-		case *DeleteSkillVersionOptions:
-			o.RegistryName = registryName
-		default:
-			return fmt.Errorf("invalid option type: %T", o)
-		}
-
-		return nil
-	}
-}
-
-// WithNamespace sets the namespace for the ListSkills, GetLatestSkillVersion,
-// or DeleteSkillVersion operation
-func WithNamespace(namespace string) Option {
-	return func(o any) error {
-		if namespace == "" {
-			return fmt.Errorf("invalid namespace: %s", namespace)
-		}
-
-		switch o := o.(type) {
-		case *ListSkillsOptions:
-			o.Namespace = namespace
-		case *GetSkillVersionOptions:
-			o.Namespace = namespace
-		case *DeleteSkillVersionOptions:
-			o.Namespace = namespace
-		default:
-			return fmt.Errorf("invalid option type: %T", o)
-		}
-
-		return nil
-	}
-}
-
-// WithNext sets the next time for the ListServerVersions operation
-func WithNext(next time.Time) Option {
-	return func(o any) error {
-		if next.IsZero() {
-			return fmt.Errorf("invalid next: %s", next)
-		}
-
-		switch o := o.(type) {
-		case *ListServerVersionsOptions:
-			o.Next = &next
-		default:
-			return fmt.Errorf("invalid option type: %T", o)
-		}
-
-		return nil
-	}
-}
-
-// WithPrev sets the prev time for the ListServerVersions operation
-func WithPrev(prev time.Time) Option {
-	return func(o any) error {
-		if prev.IsZero() {
-			return fmt.Errorf("invalid prev: %s", prev)
-		}
-
-		switch o := o.(type) {
-		case *ListServerVersionsOptions:
-			o.Prev = &prev
-		default:
-			return fmt.Errorf("invalid option type: %T", o)
-		}
-
-		return nil
-	}
-}
-
-// WithVersion sets the version for the ListServers, GetServerVersion,
-// or DeleteServerVersion operation
-func WithVersion(version string) Option {
-	return func(o any) error {
-		if version == "" {
-			return fmt.Errorf("invalid version: %s", version)
-		}
-
-		switch o := o.(type) {
-		case *ListServersOptions:
-			o.Version = version
-		case *GetServerVersionOptions:
-			o.Version = version
-		case *GetSkillVersionOptions:
-			o.Version = version
-		case *DeleteServerVersionOptions:
-			o.Version = version
-		case *DeleteSkillVersionOptions:
-			o.Version = version
-		default:
-			return fmt.Errorf("invalid option type: %T", o)
-		}
-
-		return nil
-	}
-}
-
-// WithName sets the name for the ListServerVersions, GetServerVersion,
-// or DeleteServerVersion operation
-func WithName(name string) Option {
-	return func(o any) error {
-		if name == "" {
-			return fmt.Errorf("invalid name: %s", name)
-		}
-
-		switch o := o.(type) {
-		case *ListServerVersionsOptions:
-			o.Name = name
-		case *GetServerVersionOptions:
-			o.Name = name
-		case *DeleteServerVersionOptions:
-			o.ServerName = name
-		case *ListSkillsOptions:
-			o.Name = &name
-		case *GetSkillVersionOptions:
-			o.Name = name
-		case *DeleteSkillVersionOptions:
-			o.Name = name
-		default:
-			return fmt.Errorf("invalid option type: %T", o)
-		}
-
-		return nil
-	}
-}
-
-// WithLimit sets the limit for the ListServers or ListServerVersions operation
-func WithLimit(limit int) Option {
-	return func(o any) error {
-		if limit <= 0 {
-			return fmt.Errorf("invalid limit: %d", limit)
-		}
-
-		switch o := o.(type) {
-		case *ListServersOptions:
-			o.Limit = limit
-		case *ListServerVersionsOptions:
-			o.Limit = limit
-		case *ListSkillsOptions:
-			o.Limit = limit
-		default:
-			return fmt.Errorf("invalid option type: %T", o)
-		}
-
-		return nil
-	}
-}
-
-// WithServerData sets the server data for the PublishServerVersion operation
-func WithServerData(serverData *upstreamv0.ServerJSON) Option {
-	return func(o any) error {
-		if serverData == nil {
-			return fmt.Errorf("server data is required")
-		}
-
-		switch o := o.(type) {
-		case *PublishServerVersionOptions:
-			o.ServerData = serverData
-		default:
-			return fmt.Errorf("invalid option type: %T", o)
-		}
-
-		return nil
-	}
 }
 
 // ListServersResult contains the result of a ListServers operation with pagination info.
